TLAgent/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.11.1
amqp==5.2.0
anyio==3.7.1
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
billiard==3.6.4.0
boto3==1.26.146
botocore==1.29.165
build==1.2.1
cachetools==5.4.0
celery==5.2.7
certifi==2024.7.4
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.5.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
clickhouse-connect==0.7.16
colorama==0.4.6
coloredlogs==15.0.1
cryptography==41.0.1
dataclasses-json==0.6.7
Deprecated==1.2.14
diskcache==5.6.3
distro==1.9.0
dnspython==2.6.1
duckdb==1.0.0
EbookLib==0.18
email_validator==2.2.0
et-xmlfile==1.1.0
eventlet==0.36.1
exceptiongroup==1.2.2
fastapi==0.111.1
fastapi-cli==0.0.4
filelock==3.15.4
filetype==1.2.0
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.6.1
google-ai-generativelanguage==0.2.0
google-api-core==2.19.1
google-auth==2.32.0
google-generativeai==0.1.0
googleapis-common-protos==1.63.2
greenlet==3.0.3
grpcio==1.65.1
grpcio-status==1.62.2
grpcio-tools==1.65.1
h11==0.14.0
h2==4.1.0
hnswlib==0.8.0
hpack==4.0.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.4
humanfriendly==10.0
hyperframe==6.0.1
idna==3.7
importlib_metadata==8.0.0
importlib_resources==6.4.4
intel-openmp==2021.4.0
jieba==0.42.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
json5==0.9.14
jsonpatch==1.33
jsonpointer==3.0.0
kombu==5.3.7
kubernetes==30.1.0
langchain==0.0.340
langchain-core==0.1.52
langchain-text-splitters==0.0.2
langsmith==0.0.92
llama-index==0.8.13
llama_cpp_python==0.2.82
loguru==0.7.2
lxml==5.2.2
lz4==4.3.3
Mako==1.3.5
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
mdurl==0.1.2
mkl==2021.4.0
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
msg-parser==1.2.0
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.8.1
numpy==1.26.4
oauthlib==3.2.2
olefile==0.47
onnxruntime==1.18.1
openai==1.37.1
openpyxl==3.1.5
opentelemetry-api==1.26.0
opentelemetry-exporter-otlp-proto-common==1.26.0
opentelemetry-exporter-otlp-proto-grpc==1.26.0
opentelemetry-instrumentation==0.47b0
opentelemetry-instrumentation-asgi==0.47b0
opentelemetry-instrumentation-fastapi==0.47b0
opentelemetry-proto==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
opentelemetry-util-http==0.47b0
orjson==3.10.6
overrides==7.7.0
packaging==23.2
pandas==2.2.2
pdf2image==1.17.0
pdfminer.six==20240706
pillow==10.4.0
pinecone-client==2.2.1
portalocker==2.10.1
posthog==3.5.0
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==4.25.0
psycopg2==2.9.6
pulsar-client==3.5.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==1.10.8
Pygments==2.18.0
pypandoc==1.13
PyPika==0.48.9
pyproject_hooks==1.1.0
pyreadline3==3.4.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-magic==0.4.27
python-multipart==0.0.9
python-pptx==1.0.1
pytz==2024.1
pywin32==306
PyYAML==6.0
qdrant-client==1.5.3
redis==4.5.5
regex==2024.5.15
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
rouge==1.0.1
rsa==4.9
s3transfer==0.6.2
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.0.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.16
starlette==0.37.2
sympy==1.13.1
tabulate==0.9.0
tbb==2021.11.0
tenacity==8.4.2
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.19.1
tomli==2.0.1
torch==2.3.0+cu121
torchaudio==2.3.0+cu121
torchvision==0.18.0+cu121
tqdm==4.66.4
transformers==4.41.2
typer==0.12.3
typing-inspect==0.8.0
typing_extensions==4.12.2
tzdata==2024.1
unstructured==0.8.1
urllib3==1.26.19
uvicorn==0.30.1
vine==5.1.0
watchfiles==0.22.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==12.0
win32-setctime==1.1.0
wrapt==1.16.0
xlrd==2.0.1
XlsxWriter==3.2.0
yarl==1.9.4
zipp==3.20.0
zstandard==0.23.0
